CTBase.jl

Index

CTBase.CTBase

CTBase.Control

CTBase.Controls

CTBase.Costate

CTBase.Costates

CTBase.DCostate

CTBase.DState

CTBase.Dimension

CTBase.State

CTBase.States

CTBase.Time

CTBase.Times

CTBase.TimesDisc

CTBase.Variable

CTBase.ctNumber

CTBase.ctVector

Documentation

CTBase.CTBase — Module

CTBase module.

Lists all the imported modules and packages:

Base

Core

DataStructures

DocStringExtensions

LinearAlgebra

MLStyle

Parameters

Plots

PrettyTables

Printf

ReplMaker

Unicode

List of all the exported names:

AbstractHamiltonian

AmbiguousDescription

Autonomous

BoundaryConstraint

CTCallback

CTCallbacks

CTException

Control

ControlConstraint

ControlLaw

Controls

Costate

Costates

DCostate

DState

Description

Dimension

Dynamics

FeedbackControl

Fixed

Hamiltonian

HamiltonianLift

HamiltonianVectorField

IncorrectArgument

IncorrectMethod

IncorrectOutput

Index

Lagrange

@Lie

Lie

Lift

Mayer

MixedConstraint

Model

Multiplier

NonAutonomous

NonFixed

NotImplemented

OCPInit

OptimalControlModel

OptimalControlSolution

ParsingError

Poisson

PrintCallback

State

StateConstraint

States

StopCallback

Time

TimeDependence

Times

TimesDisc

UnauthorizedCall

Variable

VariableConstraint

VariableDependence

VectorField

add

constraint

constraint!

constraint_type

constraints_labels

control!

ctNumber

ctVector

ct_repl

ctgradient

ctindices

ctinterpolate

ctjacobian

ctupperscripts

@def

dynamics!

getFullDescription

get_priority_print_callbacks

get_priority_stop_callbacks

is_max

is_min

is_time_dependent

is_time_independent

is_variable_dependent

is_variable_independent

nlp_constraints

objective!

plot

plot!

remove_constraint!

replace_call

state!

time!

variable!

∂ₜ

⋅

CTBase.Control — Type

Type alias for a control.

CTBase.Costate — Type

Type alias for an costate.

CTBase.DCostate — Type

Type alias for a tangent vector to the costate space.

CTBase.DState — Type

Type alias for a tangent vector to the state space.

CTBase.State — Type

Type alias for a state.

CTBase.TimesDisc — Type

Type alias for a grid of times.

CTBase.Variable — Type

Type alias for a variable.

CTBase.ctVector — Type

Type alias for a vector of real numbers.

CTBase.Controls — Type

Type alias for a vector of controls.

CTBase.Costates — Type

Type alias for a vector of costates.

CTBase.Dimension — Type

Type alias for a dimension.

CTBase.States — Type

Type alias for a vector of states.

CTBase.Time — Type

Type alias for a time.

CTBase.Times — Type

Type alias for a vector of times.

CTBase.ctNumber — Type

Type alias for a real number.