Functions

CTBase.BoundaryConstraint — Type

struct BoundaryConstraint{state_dimension, constraint_dimension}

Fields

f::Function

Warning

When the state is of dimension 1, consider x0 and xf as a scalar. When the constraint is dimension 1, return a scalar.

Classical calls

state dimension: 1

constraint dimension: 1

julia> B = BoundaryConstraint((t0, x0, tf, xf) -> xf - x0)
julia> B(0, 0, 1, 1)
1
julia> B(0, [0], 1, [1])
1-element Vector{Int64}:
 1

state dimension: 1

constraint dimension: 2

julia> B = BoundaryConstraint((t0, x0, tf, xf) -> [xf - x0, t0 - tf])
julia> B(0, 0, 1, 1)
2-element Vector{Int64}:
  1
 -1
julia> B(0, [0], 1, [1])
ERROR: MethodError: Cannot `convert` an object of type 
  Vector{Any} to an object of type 
  Union{Real, AbstractVector{<:Real}}

state dimension: 2

constraint dimension: 1

julia> B = BoundaryConstraint((t0, x0, tf, xf) -> tf - t0)
julia> B(0, [1, 0], 1, [0, 1])
1

state dimension: 2

constraint dimension: 2

julia> B = BoundaryConstraint((t0, x0, tf, xf) -> [tf - t0, xf[1] - x0[2]])
julia> B(0, [1, 0], 1, [1, 0])
2-element Vector{Int64}:
 1
 1

Specific calls

When giving a _Time for t0 and tf, the function takes a vector as input for x0 and xf and returns a vector as output.

Warning

To use the specific call, you must construct the function with the keywords state_dimension and constraint_dimension.

state dimension: 1

constraint dimension: 1

julia> B = BoundaryConstraint((t0, x0, tf, xf) -> xf - x0, 
    state_dimension=1, constraint_dimension=1)
julia> B(_Time(0), 0, _Time(1), 1)
ERROR: MethodError: no method matching (::BoundaryConstraint{1, 1})(::_Time, ::Int64, 
::_Time, ::Int64)
julia> B(_Time(0), [0], _Time(1), [1])
1-element Vector{Int64}:
 1

state dimension: 1

constraint dimension: 2

julia> B = BoundaryConstraint((t0, x0, tf, xf) -> [xf - x0, t0 - tf], 
    state_dimension=1, constraint_dimension=2)
julia> B(_Time(0), 0, _Time(1), 1)
ERROR: MethodError: no method matching (::BoundaryConstraint{1, 2})(::_Time, ::Int64, 
::_Time, ::Int64)
julia> B(_Time(0), [0], _Time(1), [1])
2-element Vector{Int64}:
  1
 -1

state dimension: 2

constraint dimension: 1

julia> B = BoundaryConstraint((t0, x0, tf, xf) -> tf - t0, 
    state_dimension=2, constraint_dimension=1)
julia> B(_Time(0), [1, 0], _Time(1), [0, 1])
1-element Vector{Int64}:
 1

state dimension: 2

constraint dimension: 2

julia> B = BoundaryConstraint((t0, x0, tf, xf) -> [tf - t0, xf[1] - x0[2]], 
    state_dimension=2, constraint_dimension=2)
julia> B(_Time(0), [1, 0], _Time(1), [1, 0])
2-element Vector{Int64}:
 1
 1

CTBase.ControlConstraint — Type

struct ControlConstraint{time_dependence, control_dimension, constraint_dimension}

Fields

f::Function

Similar to VectorField in the usage, but the dimension of the output of the function f is arbitrary.

CTBase.ControlLaw — Type

struct ControlLaw{time_dependence}

Fields

f::Function

The function f must be of the form f(t, x, p, args...; kwargs...) or f(x, p, args...; kwargs...) depending on the time dependence of the optimal control problem.

Note

Only classical usage.

CTBase.Dynamics — Type

struct Dynamics{time_dependence, state_dimension, control_dimension}

Fields

f::Function

Similar to Lagrange, but the function f is assumed to return a vector of the same dimension as the state x.

CTBase.FeedbackControl — Type

struct FeedbackControl{time_dependence}

Fields

f::Function

The function f must be of the form f(t, x, args...; kwargs...) or f(x, args...; kwargs...) depending on the time dependence of the optimal control problem.

Note

Only classical usage.

CTBase.Hamiltonian — Type

struct Hamiltonian{time_dependence, state_dimension}

Fields

f::Function

The default value for time_dependence is :autonomous.

Warning

When the state and adjoint are of dimension 1, consider x and p as scalars.

Classical calls

state dimension: 1

time dependence: autonomous

julia> H = Hamiltonian((x, p) -> x + p)
julia> H(1, 1)
2
julia> H([1], [1])
ERROR: MethodError: Cannot `convert` an object of type Vector{Int64} to an object of type Real
julia> H = Hamiltonian((x, p) -> x + p, time_dependence=:autonomous)
julia> H(1, 1)
2
julia> H([1], [1])
ERROR: MethodError: Cannot `convert` an object of type Vector{Int64} to an object of type Real

state dimension: 1

time dependence: nonautonomous

julia> H = Hamiltonian((t, x, p) -> x + p, time_dependence=:nonautonomous)
julia> H(1, 1, 1)
2
julia> H(1, [1], [1])
ERROR: MethodError: Cannot `convert` an object of type Vector{Int64} to an object of type Real

state dimension: 2

time dependence: autonomous

julia> H = Hamiltonian((x, p) -> x[1]^2 + p[2]^2) # or time_dependence=:autonomous
julia> H([1, 0], [0, 1])
2

state dimension: 2

time dependence: nonautonomous

julia> H = Hamiltonian((t, x, p) -> t + x[1]^2 + p[2]^2, time_dependence=:nonautonomous)
julia> H(1, [1, 0], [0, 1])
3

Specific calls

When giving a _Time for t, the function takes a vector as input for x and p.

state dimension: 1

time dependence: autonomous

julia> H = Hamiltonian((x, p) -> x + p, state_dimension=1)
julia> H(_Time(0), 1, 1)
ERROR: MethodError: no method matching (::Hamiltonian{:autonomous, 1})(::CTBase._Time, ::Int64, ::Int64)
julia> H(_Time(0), [1], [1])
2

state dimension: 1

time dependence: nonautonomous

julia> H = Hamiltonian((t, x, p) -> t + x + p, state_dimension=1, time_dependence=:nonautonomous)
julia> H(_Time(1), 1, 1)
ERROR: MethodError: no method matching (::Hamiltonian{:nonautonomous, 1})(::CTBase._Time, ::Int64, ::Int64)
julia> H(_Time(1), [1], [1])
3

state dimension: 2

time dependence: autonomous

julia> H = Hamiltonian((x, p) -> x[1]^2 + p[2]^2, state_dimension=2)
julia> H(_Time(0), [1, 0], [0, 1])
2

state dimension: 2

time dependence: nonautonomous

julia> H = Hamiltonian((t, x, p) -> t + x[1]^2 + p[2]^2, state_dimension=2, time_dependence=:nonautonomous)
julia> H(_Time(1), [1, 0], [0, 1])
3

CTBase.HamiltonianVectorField — Type

struct HamiltonianVectorField{time_dependence, state_dimension}

Fields

f::Function

The default value for time_dependence is :autonomous.

Warning

When the state and adjoint are of dimension 1, consider x and p as scalars.

Classical calls

state dimension: 1

time dependence: autonomous

julia> Hv = HamiltonianVectorField((x, p) -> [x + p, x - p])
julia> Hv(1, 1)
2-element Vector{Int64}:
 2
 0
julia> Hv([1], [1])
ERROR: MethodError: Cannot `convert` an object of type 
  Vector{Vector{Int64}} to an object of type 
  AbstractVector{<:Real}
julia> Hv = HamiltonianVectorField((x, p) -> [x + p, x - p], time_dependence=:autonomous)
julia> Hv(1, 1)
2-element Vector{Int64}:
 2
 0
julia> Hv([1], [1])
ERROR: MethodError: Cannot `convert` an object of type 
  Vector{Vector{Int64}} to an object of type 
  AbstractVector{<:Real}

state dimension: 1

time dependence: nonautonomous

julia> Hv = HamiltonianVectorField((t, x, p) -> [x + p, x - p], time_dependence=:nonautonomous)
julia> Hv(1, 1, 1)
2-element Vector{Int64}:
 2
 0
julia> Hv(1, [1], [1])
ERROR: MethodError: Cannot `convert` an object of type 
  Vector{Vector{Int64}} to an object of type 
  AbstractVector{<:Real}

state dimension: 2

time dependence: autonomous

julia> Hv = HamiltonianVectorField((x, p) -> [x[1]^2 + p[2]^2, x[2]^2 - p[1]^2]) # or time_dependence=:autonomous
julia> Hv([1, 0], [0, 1])
2-element Vector{Int64}:
 2
 0

state dimension: 2

time dependence: nonautonomous

julia> Hv = HamiltonianVectorField((t, x, p) -> [t + x[1]^2 + p[2]^2, x[2]^2 - p[1]^2], time_dependence=:nonautonomous)
julia> Hv(1, [1, 0], [0, 1])
2-element Vector{Int64}:
 3
 0

Specific calls

When giving a _Time for t, the function takes a vector as input for x and p.

state dimension: 1

time dependence: autonomous

julia> Hv = HamiltonianVectorField((x, p) -> [x + p, x - p], state_dimension=1)
julia> Hv(_Time(0), 1, 1)
ERROR: MethodError: no method matching (::HamiltonianVectorField{:autonomous, 1})(::CTBase._Time, ::Int64, ::Int64)
julia> Hv(_Time(0), [1], [1])
2-element Vector{Int64}:
 2
 0

state dimension: 1

time dependence: nonautonomous

julia> Hv = HamiltonianVectorField((t, x, p) -> [t + x + p, x - p], state_dimension=1, time_dependence=:nonautonomous)
julia> Hv(_Time(1), 1, 1)
ERROR: MethodError: no method matching (::HamiltonianVectorField{:nonautonomous, 1})(::CTBase._Time, ::Int64, ::Int64)
julia> Hv(_Time(1), [1], [1])
2-element Vector{Int64}:
 3
 0

state dimension: 2

time dependence: autonomous

julia> Hv = HamiltonianVectorField((x, p) -> [x[1]^2 + p[2]^2, x[2]^2 - p[1]^2], state_dimension=2)
julia> Hv(_Time(0), [1, 0], [0, 1])
2-element Vector{Int64}:
 2
 0

state dimension: 2

time dependence: nonautonomous

julia> Hv = HamiltonianVectorField((t, x, p) -> [t + x[1]^2 + p[2]^2, x[2]^2 - p[1]^2], state_dimension=2, time_dependence=:nonautonomous)
julia> Hv(_Time(1), [1, 0], [0, 1])
2-element Vector{Int64}:
 3
 0

CTBase.Lagrange — Type

struct Lagrange{time_dependence, state_dimension, control_dimension}

Fields

f::Function

The default value for time_dependence is :autonomous.

Warning

When the state is of dimension 1, consider x as a scalar. Same for the control.

Classical calls

state dimension: 1

control dimension: 1

time dependence: autonomous

julia> L = Lagrange((x, u) -> x + u) # or time_dependence=:autonomous
julia> L(1, 1)
2
julia> L([1], [1])
ERROR: MethodError: Cannot `convert` an object of type Vector{Int64} to an object of type Real

state dimension: 1

control dimension: 1

time dependence: nonautonomous

julia> L = Lagrange((t, x, u) -> x + u, time_dependence=:nonautonomous)
julia> L(1, 1, 1)
2
julia> L(1, [1], [1])
ERROR: MethodError: Cannot `convert` an object of type Vector{Int64} to an object of type Real

state dimension: 2

control dimension: 1

time dependence: autonomous

julia> L = Lagrange((x, u) -> x[1]^2 + u^2) # or time_dependence=:autonomous
julia> L([1, 2], 1)
2

state dimension: 2

control dimension: 1

time dependence: nonautonomous

julia> L = Lagrange((t, x, u) -> x[1]^2 + u^2, time_dependence=:nonautonomous)
julia> L(1, [1, 2], 1)
2

state dimension: 1

control dimension: 2

time dependence: autonomous

julia> L = Lagrange((x, u) -> x^2 + u[2]^2) # or time_dependence=:autonomous
julia> L(1, [1, 2])
5

state dimension: 1

control dimension: 2

time dependence: nonautonomous

julia> L = Lagrange((t, x, u) -> t + x^2 + u[2]^2, time_dependence=:nonautonomous)
julia> L(1, 1, [1, 2])
6

state dimension: 2

control dimension: 2

time dependence: autonomous

julia> L = Lagrange((x, u) -> x[1]^2 + u[2]^2) # or time_dependence=:autonomous
julia> L([1, 2], [1, 2])
5

state dimension: 2

control dimension: 2

time dependence: nonautonomous

julia> L = Lagrange((t, x, u) -> t + x[1]^2 + u[2]^2, time_dependence=:nonautonomous)
julia> L(1, [1, 2], [1, 2])
6

Specific calls

state dimension: 1

control dimension: 1

time dependence: autonomous

julia> L = Lagrange((x, u) -> x + u, state_dimension=1, control_dimension=1) # or time_dependence=:autonomous
julia> L(_Time(1), 1, 1)
ERROR: MethodError: no method matching (::Lagrange{:autonomous, 0, 0})(::CTBase._Time, ::Int64, ::Int64)
julia> L(_Time(1), [1], [1])
2

state dimension: 1

control dimension: 1

time dependence: nonautonomous

julia> L = Lagrange((t, x, u) -> t + x + u, state_dimension=1, control_dimension=1, time_dependence=:nonautonomous)
julia> L(_Time(1), 1, 1)
ERROR: MethodError: no method matching (::Lagrange{:nonautonomous, 0, 0})(::CTBase._Time, ::Int64, ::Int64)
julia> L(_Time(1), [1], [1])
3

state dimension: 2

control dimension: 1

time dependence: autonomous

julia> L = Lagrange((x, u) -> x[1]^2 + u^2, state_dimension=2, control_dimension=1) # or time_dependence=:autonomous
julia> L(_Time(1), [1, 2], 1)
ERROR: MethodError: no method matching (::Lagrange{:autonomous, 0, 0})(::CTBase._Time, ::Vector{Int64}, ::Int64)
julia> L(_Time(1), [1, 2], [1])
2

state dimension: 2

control dimension: 1

time dependence: nonautonomous

julia> L = Lagrange((t, x, u) -> t + x[1]^2 + u^2, state_dimension=2, control_dimension=1, time_dependence=:nonautonomous)
julia> L(_Time(1), [1, 2], 1)
ERROR: MethodError: no method matching (::Lagrange{:nonautonomous, 0, 0})(::CTBase._Time, ::Vector{Int64}, ::Int64)
julia> L(_Time(1), [1, 2], [1])
3

state dimension: 1

control dimension: 2

time dependence: autonomous

julia> L = Lagrange((x, u) -> x^2 + u[2]^2, state_dimension=1, control_dimension=2) # or time_dependence=:autonomous
julia> L(_Time(1), 1, [1, 2])
ERROR: MethodError: no method matching (::Lagrange{:autonomous, 0, 0})(::CTBase._Time, ::Int64, ::Vector{Int64})
julia> L(_Time(1), [1], [1, 2])
5

state dimension: 1

control dimension: 2

time dependence: nonautonomous

julia> L = Lagrange((t, x, u) -> t + x^2 + u[2]^2, state_dimension=1, control_dimension=2, time_dependence=:nonautonomous)
julia> L(_Time(1), 1, [1, 2])
ERROR: MethodError: no method matching (::Lagrange{:nonautonomous, 0, 0})(::CTBase._Time, ::Int64, ::Vector{Int64})
julia> L(_Time(1), [1], [1, 2])
6

state dimension: 2

control dimension: 2

time dependence: autonomous

julia> L = Lagrange((x, u) -> x[1]^2 + u[2]^2, state_dimension=2, control_dimension=2) # or time_dependence=:autonomous
julia> L(_Time(1), [1, 2], [1, 2])
5

CTBase.Mayer — Type

struct Mayer{state_dimension}

Fields

f::Function

Warning

When the state is of dimension 1, consider x0 and xf as a scalar.

Classical calls

state dimension: 1

julia> G = Mayer((t0, x0, tf, xf) -> xf - x0)
julia> G(0, 0, 1, 1)
1
julia> G(0, [0], 1, [1])
ERROR: MethodError: Cannot `convert` an object of type Vector{Int64} to an object of type Real

state dimension: 2

julia> G = Mayer((t0, x0, tf, xf) -> tf - t0)
julia> G(0, [1, 0], 1, [0, 1])
1

Specific calls

When giving a _Time for t0 and tf, the function takes a vector as input for x0 and xf.

Warning

To use the specific call, you must construct the function with the keyword state_dimension.

state dimension: 1

julia> G = Mayer((t0, x0, tf, xf) -> xf - x0, state_dimension=1)
julia> G(_Time(0), 0, _Time(1), 1)
ERROR: MethodError: no method matching (:Mayer{1})(::CTBase._Time, ::Int64, ::CTBase._Time, ::Int64)
julia> G(_Time(0), [0], _Time(1), [1])
1

state dimension: 2

julia> G = Mayer((t0, x0, tf, xf) -> tf - t0, state_dimension=2)
julia> G(_Time(0), [1, 0], _Time(1), [0, 1])
1

CTBase.MixedConstraint — Type

struct MixedConstraint{time_dependence, state_dimension, control_dimension, constraint_dimension}

Fields

f::Function

Similar to Lagrange in the usage, but the dimension of the output of the function f is arbitrary.

CTBase.Multiplier — Type

struct Multiplier{time_dependence}

Fields

f::Function

The function f must be of the form f(t, x, p, args...; kwargs...) or f(x, p, args...; kwargs...) depending on the time dependence of the optimal control problem.

Note

Only classical usage.

CTBase.StateConstraint — Type

struct StateConstraint{time_dependence, state_dimension, constraint_dimension}

Fields

f::Function

Similar to VectorField in the usage, but the dimension of the output of the function f is arbitrary.

CTBase.VectorField — Type

struct VectorField{time_dependence, state_dimension}

Fields

f::Function

The default value for time_dependence is :autonomous.

Warning

When the state is of dimension 1, consider x as a scalar.

Classical calls

state dimension: 1

time dependence: autonomous

julia> V = VectorField(x -> 2x) # or time_dependence=:autonomous
julia> V(1)
2
julia> V([1])
1-element Vector{Int64}:
 2

state dimension: 1

time dependence: nonautonomous

julia> V = VectorField((t, x) -> t+2x, time_dependence=:nonautonomous)
julia> V(1, 1)
3

state dimension: 2

time dependence: autonomous

julia> V = VectorField(x -> [x[1]^2, x[2]^2]) # or time_dependence=:autonomous
julia> V([1, 0])
2-element Vector{Int64}:
 1
 0

state dimension: 2

time dependence: nonautonomous

julia> V = VectorField((t, x) -> [t + x[1]^2, x[2]^2], time_dependence=:nonautonomous)
julia> V(1, [1, 0])
2-element Vector{Int64}:
 2
 0

Specific calls

state dimension: 1

time dependence: autonomous

julia> V = VectorField(x -> 2x, state_dimension=1) # or time_dependence=:autonomous
julia> V(_Time(0), 1)
ERROR: MethodError: no method matching (::VectorField{:autonomous, 1})(::CTBase._Time, ::Int64)
julia> V(_Time(0), [1])
1-element Vector{Int64}:
 2

state dimension: 1

time dependence: nonautonomous

julia> V = VectorField((t, x) -> t+2x, state_dimension=1, time_dependence=:nonautonomous)
julia> V(_Time(0), 1)
ERROR: MethodError: no method matching (::VectorField{:nonautonomous, 1})(::CTBase._Time, ::Int64)
julia> V(_Time(1), [1])
1-element Vector{Int64}:
 3

state dimension: N>1

time dependence: autonomous

julia> V = VectorField(x -> [x[1]^2, -x[2]^2], state_dimension=2) # or time_dependence=:autonomous
julia> V(_Time(0), [1, 2])
2-element Vector{Int64}:
  1
 -4

state dimension: N>1

time dependence: nonautonomous

julia> V = VectorField((t, x) -> [t + x[1]^2, -x[2]^2], state_dimension=2, time_dependence=:nonautonomous)
julia> V(_Time(1), [1, 2])
2-element Vector{Int64}:
  2
 -4

CTBase._Time — Type

struct _Time

Fields

value::Real

Used to define typed methods to make specific calls.