Model

CTBase.Index — Type

mutable struct Index

Fields

val::Integer

CTBase.OptimalControlModel — Type

mutable struct OptimalControlModel{time_dependence} <: CTBase.AbstractOptimalControlModel

Fields

initial_time::Union{Nothing, Real}: Default: nothing

final_time::Union{Nothing, Real}: Default: nothing

time_name::Union{Nothing, String}: Default: nothing

lagrange::Union{Nothing, Lagrange}: Default: nothing

mayer::Union{Nothing, Mayer}: Default: nothing

criterion::Union{Nothing, Symbol}: Default: nothing

dynamics::Union{Nothing, Dynamics}: Default: nothing

state_dimension::Union{Nothing, Integer}: Default: nothing

state_names::Union{Nothing, Vector{String}}: Default: nothing

control_dimension::Union{Nothing, Integer}: Default: nothing

control_names::Union{Nothing, Vector{String}}: Default: nothing

constraints::Dict{Symbol, Tuple}: Default: Dict{Symbol, Tuple{Vararg{Any}}}()

CTBase.Model — Method

Model(; time_dependence) -> OptimalControlModel{:autonomous}

Return a new OptimalControlModel instance, that is a model of an optimal control problem.

The model is defined by the following optional keyword argument:

time_dependence: either :autonomous or :nonautonomous. Default is :autonomous.

Examples

julia> ocp = Model()
julia> ocp = Model(time_dependence=:nonautonomous)

Note

If the time dependence of the model is defined as nonautonomous, then, the dynamics function, the lagrange cost and the path constraints must be defined as functions of time and state, and possibly control. If the model is defined as autonomous, then, the dynamics function, the lagrange cost and the path constraints must be defined as functions of state, and possibly control.

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    lb,
    ub
) -> Tuple{Symbol, Symbol, CTBase.var"#145#146", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#147#148", Any, Any}

Add an :initial, :final, :control or :state box constraint (whole range).

Examples

julia> constraint!(ocp, :initial, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :final, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :control, [ 0, 0 ], [ 2, 3 ])
julia> constraint!(ocp, :state, [ 0, 0, 0 ], [ 1, 2, 1 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    val
) -> Any
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#145#146", Any, Any}

Add an :initial or :final value constraint on the state.

Examples

# state dimension: 1
julia> constraint!(ocp, :initial, 0)
julia> constraint!(ocp, :final, 1)

# state dimension: 2
julia> constraint!(ocp, :initial, [ 0, 0])
julia> constraint!(ocp, :final, [ 0, 0])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    val
) -> Tuple{Symbol, Symbol, Function, Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, Function, Any, Any}

Add a :boundary, :control, :state or :mixed value functional constraint.

Examples

julia> constraint!(ocp, :boundary, (t0, x0, tf, xf) -> [t0+tf, x0[3]+xf[2]], [ 0, 0 ])

# autonomous ocp
julia> constraint!(ocp, :control, u -> 2u, 1)
julia> constraint!(ocp, :state, x -> x-1, [ 0, 0, 0])
julia> constraint!(ocp, :mixed, (x, u) -> x[1]-u, 0)

# nonautonomous ocp
julia> constraint!(ocp, :control, (t, u) -> 2u, 1)
julia> constraint!(ocp, :state, (t, x) -> x-t, [ 0, 0, 0])
julia> constraint!(ocp, :mixed, (t, x, u) -> x[1]-u, 0)

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    lb,
    ub
) -> Tuple{Symbol, Symbol, Function, Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, Function, Any, Any}

Add a :boundary, :control, :state or :mixed box functional constraint.

Examples

julia> constraint!(ocp, :boundary, (t0, x0, tf, xf) -> [t0+tf, x0[3]+xf[2]], [ 0, 0 ], [ 1, 1 ])

# autonomous ocp
julia> constraint!(ocp, :control, u -> 2u, 0, 1)
julia> constraint!(ocp, :state, x -> x-1, [ 0, 0, 0], [ 1, 2, 1 ])
julia> constraint!(ocp, :mixed, (x, u) -> x[1]-u, 0, 1)

# nonautonomous ocp
julia> constraint!(ocp, :control, (t, u) -> 2u, 0, 1)
julia> constraint!(ocp, :state, (t, x) -> x-t, [ 0, 0, 0], [ 1, 2, 1 ])
julia> constraint!(ocp, :mixed, (t, x, u) -> x[1]-u, 0, 1)

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, OrdinalRange{<:Integer}},
    val
) -> Tuple{Symbol, Symbol, CTBase.var"#147#148", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, OrdinalRange{<:Integer}},
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#147#148", Any, Any}

Add an :initial or :final value constraint on a range of the state.

Examples

julia> constraint!(ocp, :initial, 1:2:5, [ 0, 0, 0 ])
julia> constraint!(ocp, :initial, 2:3, [ 0, 0 ])
julia> constraint!(ocp, :final, Index(2), 0)

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, OrdinalRange{<:Integer}},
    lb,
    ub
) -> Tuple{Symbol, Symbol, CTBase.var"#147#148", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, OrdinalRange{<:Integer}},
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, Any, Any, Any}

Add an :initial, :final, :control or :state box constraint on a range.

Examples

julia> constraint!(ocp, :initial, 2:3, [ 0, 0 ], [1, 2])
julia> constraint!(ocp, :final, Index(1), 0, 2)
julia> constraint!(ocp, :control, Index(1), 0, 2)
julia> constraint!(ocp, :state, 2:3, [ 0, 0 ], [1, 2])
julia> constraint!(ocp, :initial, 1:2:5, [ 0, 0, 0], [1, 2, 1])

CTBase.constraint! — Method

constraint!(
    ocp::OptimalControlModel{time_dependence},
    type::Symbol,
    f::Function
) -> Any

Set the dynamics f(x, u) in the autonomous case and f(t, x, u) in the nonautonomous case.

Note

The dimensions of the state and control must be set before adding the dynamics.

The dynamics must be provided as a function of time (if ocp is nonautonomous), state and control.

The output of the dynamics must be of same dimension as the state (scalar if the state is of dimension 1).

Example

julia> constraint!(ocp, :dynamics, f)

CTBase.constraint — Method

constraint(
    ocp::OptimalControlModel{:autonomous},
    label::Symbol
) -> Function

Retrieve a labeled constraint. The result is a function associated with the constraint computation (not taking into account provided value / bounds).

Example

julia> constraint!(ocp, :initial, 0, :c0)
julia> c = constraint(ocp, :c0)
julia> c(1)
1

CTBase.constraints_labels — Method

constraints_labels(
    ocp::OptimalControlModel
) -> Base.KeySet{Symbol, Dict{Symbol, Tuple}}

Return the labels of the constraints as a Base.keys.

Example

julia> constraints_labels(ocp)

CTBase.control! — Function

control!(ocp::OptimalControlModel, m::Integer)
control!(
    ocp::OptimalControlModel,
    m::Integer,
    names::Union{String, Vector{String}}
)

Define the control dimension and possibly the names of each coordinate.

Examples

julia> control!(ocp, 1)
julia> ocp.control_dimension
1
julia> ocp.control_names
["u"]

julia> control!(ocp, 1, "v")
julia> ocp.control_dimension
1
julia> ocp.control_names
["v"]

julia> control!(ocp, 2)
julia> ocp.control_dimension
2
julia> ocp.control_names
["u₁", "u₂"]

julia> control!(ocp, 2, [ "v₁", "v₂" ])
julia> ocp.control_dimension
2
julia> ocp.control_names
["v₁", "v₂"]

julia> control!(ocp, 2, :v)
julia> ocp.control_dimension
2
julia> ocp.control_names
["v₁", "v₂"]

julia> control!(ocp, 2, "v")
julia> ocp.control_dimension
2
julia> ocp.control_names
["v₁", "v₂"]

CTBase.isautonomous — Method

isautonomous(ocp::OptimalControlModel) -> Bool

Return true if the model has been defined as autonomous.

CTBase.isautonomous — Method

isautonomous(time_dependence::Symbol) -> Bool

Return !isnonautonomous(time_dependence)

CTBase.ismax — Method

ismax(ocp::OptimalControlModel) -> Bool

Return true if the criterion type of ocp is :max.

CTBase.ismin — Method

ismin(ocp::OptimalControlModel) -> Bool

Return true if the criterion type of ocp is :min.

CTBase.isnonautonomous — Method

isnonautonomous(time_dependence::Symbol) -> Bool

Return :nonautonomous == time_dependence

CTBase.isnonautonomous — Method

isnonautonomous(
    ocp::OptimalControlModel{time_dependence}
) -> Bool

Return true if the model has been defined as nonautonomous.

CTBase.nlp_constraints — Method

nlp_constraints(
    ocp::OptimalControlModel{time_dependence}
) -> Tuple{Tuple{Vector{Real}, CTBase.var"#ξ#167", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#η#168", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#ψ#169", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#ϕ#170", Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}}

Return a 6-tuple of tuples:

(ξl, ξ, ξu) are control constraints

(ηl, η, ηu) are state constraints

(ψl, ψ, ψu) are mixed constraints

(ϕl, ϕ, ϕu) are boundary constraints

(ulb, uind, uub) are control linear constraints of a subset of indices

(xlb, xind, xub) are state linear constraints of a subset of indices

Note

The dimensions of the state and control must be set before calling nlp_constraints.

Example

julia> (ξl, ξ, ξu), (ηl, η, ηu), (ψl, ψ, ψu), (ϕl, ϕ, ϕu), 
    (ulb, uind, uub), (xlb, xind, xub) = nlp_constraints(ocp)

CTBase.objective! — Method

objective!(
    ocp::OptimalControlModel{time_dependence},
    type::Symbol,
    g::Function,
    f⁰::Function
) -> Lagrange
objective!(
    ocp::OptimalControlModel{time_dependence},
    type::Symbol,
    g::Function,
    f⁰::Function,
    criterion::Symbol
) -> Lagrange

Set the criterion to the function g and f⁰. Type can be :bolza. Criterion is :min or :max.

Note

The dimensions of the state and control must be set before adding the dynamics.

Example

julia> objective!(ocp, :bolza, (t0, x0, tf, xf) -> tf, (x, u) -> x[1]^2 + u[1]^2) 

CTBase.objective! — Method

objective!(
    ocp::OptimalControlModel{time_dependence},
    type::Symbol,
    f::Function
) -> Any
objective!(
    ocp::OptimalControlModel{time_dependence},
    type::Symbol,
    f::Function,
    criterion::Symbol
) -> Any

Set the criterion to the function f. Type can be :mayer or :lagrange. Criterion is :min or :max.

Note

The dimensions of the state and control must be set before adding the dynamics.

Examples

julia> objective!(ocp, :mayer, (t0, x0, tf, xf) -> tf)
julia> objective!(ocp, :lagrange, (x, u) -> x[1]^2 + u[1]^2)

Warning

If you set twice the objective, only the last one will be taken into account.

CTBase.remove_constraint! — Method

remove_constraint!(
    ocp::OptimalControlModel,
    label::Symbol
) -> Dict{Symbol, Tuple}

Remove a labeled constraint.

Example

julia> remove_constraint!(ocp, :con)

CTBase.state! — Function

state!(ocp::OptimalControlModel, n::Integer)
state!(
    ocp::OptimalControlModel,
    n::Integer,
    names::Union{String, Vector{String}}
)

Define the state dimension and possibly the names of each coordinate.

Examples

julia> state!(ocp, 1)
julia> ocp.state_dimension
1
julia> ocp.state_names
["x"]

julia> state!(ocp, 1, "y")
julia> ocp.state_dimension
1
julia> ocp.state_names
["y"]

julia> state!(ocp, 2)
julia> ocp.state_dimension
2
julia> ocp.state_names
["x₁", "x₂"]

julia> state!(ocp, 2, [ "y₁", "y₂" ])
julia> ocp.state_dimension
2
julia> ocp.state_names
["y₁", "y₂"]

julia> state!(ocp, 2, :y)
julia> ocp.state_dimension
2
julia> ocp.state_names
["y₁", "y₂"]

julia> state!(ocp, 2, "y")
julia> ocp.state_dimension
2
julia> ocp.state_names
["y₁", "y₂"]

CTBase.time! — Function

time!(
    ocp::OptimalControlModel,
    times::AbstractVector{<:Real}
)
time!(
    ocp::OptimalControlModel,
    times::AbstractVector{<:Real},
    name::String
)

Fix initial and final times to times[1] and times[2], respectively.

Note

You can use time! once to set either the initial or the final time, or both.

Examples

julia> time!(ocp, [0, 1])
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"t"

julia> time!(ocp, [0, 1], "s")
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"s"

julia> time!(ocp, [0, 1], :s)
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"s"

CTBase.time! — Function

time!(ocp::OptimalControlModel, type::Symbol, time::Real)
time!(
    ocp::OptimalControlModel,
    type::Symbol,
    time::Real,
    name::String
)

Fix initial (resp. final) time, the final (resp. initial) time being variable (free), when type is :initial. And conversely when type is :final.

Note

You can use time! once to set either the initial or the final time, or both.

Examples

julia> time!(ocp, :initial, 0)
julia> ocp.initial_time
0
julia> ocp.time_name
"t"

julia> time!(ocp, :initial, 0, "s")
julia> ocp.initial_time
0
julia> ocp.time_name
"s"

julia> time!(ocp, :initial, 0, :s)
julia> ocp.initial_time
0
julia> ocp.time_name
"s"

julia> time!(ocp, :final, 1)
julia> ocp.final_time
1
julia> ocp.time_name
"t"

julia> time!(ocp, :final, 1, "s")
julia> ocp.final_time
1
julia> ocp.time_name
"s"

julia> time!(ocp, :final, 1, :s)
julia> ocp.final_time
1
julia> ocp.time_name
"s"