Description

CTBase.Description — Type

A description is a tuple of symbols, that is a Tuple{Vararg{Symbol}}.

Base.:\ — Method

\(
    x::Tuple{Vararg{Symbol}},
    y::Tuple{Vararg{Symbol}}
) -> Tuple{Vararg{Symbol}}

Returns the difference between the description x and the description y.

Example

julia> (:a, :b) \ (:a,)
(:b,)

CTBase.add — Method

add(
    x::Tuple{Vararg{Tuple{Vararg{Symbol}}}},
    y::Tuple{Vararg{Symbol}}
) -> Tuple{Tuple{Vararg{Symbol}}}

Concatenate the description y at the tuple of descriptions x if it is not already in the tuple x.

Example

julia> descriptions = ()
julia> descriptions = add(descriptions, (:a,))
((:a,),)
julia> descriptions = add(descriptions, (:b,))
((:a,), (:b,))

CTBase.add — Method

add(
    x::Tuple{},
    y::Tuple{Vararg{Symbol}}
) -> Tuple{Tuple{Vararg{Symbol}}}

Returns a tuple containing only the description y.

Example

julia> descriptions = ()
julia> descriptions = add(descriptions, (:a,))
((:a,),)
julia> descriptions[1]
(:a,)

CTBase.getFullDescription — Method

getFullDescription(
    desc::Tuple{Vararg{Symbol}},
    desc_list::Tuple{Vararg{Tuple{Vararg{Symbol}}}}
) -> Tuple{Vararg{Symbol}}

Returns a complete description from an incomplete description desc and a list of complete descriptions desc_list. If several complete descriptions are possible, then the first one is returned.

Example

julia> desc_list = ((:a, :b), (:b, :c), (:a, :c))
((:a, :b), (:b, :c), (:a, :c))
julia> getFullDescription((:a,), desc_list)
(:a, :b)

CTBase.makeDescription — Method

makeDescription(
    desc::Tuple{Vararg{Symbol}}
) -> Tuple{Vararg{Symbol}}

Returns the description.

Example

julia> makeDescription((:a, :b))
(:a, :b)

CTBase.makeDescription — Method

makeDescription(desc::Symbol...) -> Tuple{Vararg{Symbol}}

Create a description from a tuple of symbols.

Example

julia> makeDescription(:a, :b)
(:a, :b)