Model

CTBase.Model — Method

Model(
    dependencies::DataType...
) -> OptimalControlModel{Autonomous, Fixed}

Return a new OptimalControlModel instance, that is a model of an optimal control problem.

The model is defined by the following argument:

dependencies: either Autonomous or NonAutonomous. Default is Autonomous. And either NonFixed or Fixed. Default is Fixed.

Examples

julia> ocp = Model()
julia> ocp = Model(NonAutonomous)
julia> ocp = Model(Autonomous, NonFixed)

Note

If the time dependence of the model is defined as nonautonomous, then, the dynamics function, the lagrange cost and the path constraints must be defined as functions of time and state, and possibly control. If the model is defined as autonomous, then, the dynamics function, the lagrange cost and the path constraints must be defined as functions of state, and possibly control.

CTBase.Model — Method

Model(
;
    autonomous,
    variable
) -> OptimalControlModel{Autonomous, Fixed}

Return a new OptimalControlModel instance, that is a model of an optimal control problem.

The model is defined by the following optional keyword argument:

autonomous: either true or false. Default is true.

variable: either true or false. Default is false.

Examples

julia> ocp = Model()
julia> ocp = Model(autonomous=false)
julia> ocp = Model(autonomous=false, variable=true)

Note

If the time dependence of the model is defined as nonautonomous, then, the dynamics function, the lagrange cost and the path constraints must be defined as functions of time and state, and possibly control. If the model is defined as autonomous, then, the dynamics function, the lagrange cost and the path constraints must be defined as functions of state, and possibly control.

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    val::Union{Real, AbstractVector{<:Real}}
)
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    val::Union{Real, AbstractVector{<:Real}},
    label::Symbol
)

Add a :boundary, :control, :state, :mixed or :variable value functional constraint.

Note

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

When an element is of dimension 1, consider it as a scalar.

Examples

# variable independent ocp
julia> constraint!(ocp, :boundary, (x0, xf) -> x0[3]+xf[2], 0)

# variable dependent ocp
julia> constraint!(ocp, :boundary, (x0, xf, v) -> x0[3]+xf[2]*v[1], 0)

# time independent and variable independent ocp
julia> constraint!(ocp, :control, u -> 2u, 1)
julia> constraint!(ocp, :state, x -> x-1, [ 0, 0, 0 ])
julia> constraint!(ocp, :mixed, (x, u) -> x[1]-u, 0)

# time dependent and variable independent ocp
julia> constraint!(ocp, :control, (t, u) -> 2u, 1)
julia> constraint!(ocp, :state, (t, x) -> x-t, [ 0, 0, 0 ])
julia> constraint!(ocp, :mixed, (t, x, u) -> x[1]-u, 0)

# time independent and variable dependent ocp
julia> constraint!(ocp, :control, (u, v) -> 2u*v[1], 1)
julia> constraint!(ocp, :state, (x, v) -> x-v[2], [ 0, 0, 0 ])
julia> constraint!(ocp, :mixed, (x, u) -> x[1]-u+v[1], 0)

# time dependent and variable dependent ocp
julia> constraint!(ocp, :control, (t, u, v) -> 2u-t*v[2], 1)
julia> constraint!(ocp, :state, (t, x, v) -> x-t+v[1], [ 0, 0, 0 ])
julia> constraint!(ocp, :mixed, (t, x, u, v) -> x[1]-u*v[1], 0)

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    val::Union{Real, AbstractVector{<:Real}}
)
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    val::Union{Real, AbstractVector{<:Real}},
    label::Symbol
)

Add an :initial or :final value constraint on the state, or a :variable value. Can also be used with :state and :control.

Note

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

When an element is of dimension 1, consider it as a scalar.

Examples

julia> constraint!(ocp, :initial, [ 0, 0 ])
julia> constraint!(ocp, :final, 2) # if the state is of dimension 1
julia> constraint!(ocp, :variable, [ 3, 0, 1 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    lb::Union{Real, AbstractVector{<:Real}},
    ub::Union{Real, AbstractVector{<:Real}}
)
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    lb::Union{Real, AbstractVector{<:Real}},
    ub::Union{Real, AbstractVector{<:Real}},
    label::Symbol
)

Add an :initial, :final, :control, :state or :variable box constraint (whole range).

Note

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

When an element is of dimension 1, consider it as a scalar.

Examples

julia> constraint!(ocp, :initial, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :final, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :control, [ 0, 0 ], [ 2, 3 ])
julia> constraint!(ocp, :state, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :variable, 0, 1) # the variable here is of dimension 1

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, OrdinalRange{<:Integer}},
    val::Union{Real, AbstractVector{<:Real}}
)
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, OrdinalRange{<:Integer}},
    val::Union{Real, AbstractVector{<:Real}},
    label::Symbol
)

Add an :initial or :final value constraint on a range of the state, or a value constraint on a range of the :variable.

Note

The range of the constraint must be contained in 1:n if the constraint is on the state, or 1:q if the constraint is on the variable.

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

Examples

julia> constraint!(ocp, :initial, 1:2:5, [ 0, 0, 0 ])
julia> constraint!(ocp, :initial, 2:3, [ 0, 0 ])
julia> constraint!(ocp, :final, Index(2), 0)
julia> constraint!(ocp, :variable, 2:3, [ 0, 3 ])

CTBase.constraint! — Method

constraint!(
    ocp::OptimalControlModel,
    type::Symbol;
    rg,
    f,
    val,
    lb,
    ub,
    label
)

Add an :initial, :final, :control, :state or :variable box constraint on a range.

Note

The range of the constraint must be contained in 1:n if the constraint is on the state, or 1:m if the constraint is on the control, or 1:q if the constraint is on the variable.

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

Examples

julia> constraint!(ocp, :initial, rg=2:3, lb=[ 0, 0 ], ub=[ 1, 2 ])
julia> constraint!(ocp, :final, val=Index(1), lb=0, ub=2)
julia> constraint!(ocp, :control, val=Index(1), lb=0, ub=2)
julia> constraint!(ocp, :state, rg=2:3, lb=[ 0, 0 ], ub=[ 1, 2 ])
julia> constraint!(ocp, :initial, rg=1:2:5, lb=[ 0, 0, 0 ], ub=[ 1, 2, 1 ])
julia> constraint!(ocp, :variable, rg=1:2, lb=[ 0, 0 ], ub=[ 1, 2 ])

CTBase.constraint! — Method

constraint!(
    ocp::OptimalControlModel{<:TimeDependence, V<:VariableDependence},
    type::Symbol,
    rg::Union{Index, OrdinalRange{<:Integer}},
    lb::Union{Real, AbstractVector{<:Real}},
    ub::Union{Real, AbstractVector{<:Real}}
)
constraint!(
    ocp::OptimalControlModel{<:TimeDependence, V<:VariableDependence},
    type::Symbol,
    rg::Union{Index, OrdinalRange{<:Integer}},
    lb::Union{Real, AbstractVector{<:Real}},
    ub::Union{Real, AbstractVector{<:Real}},
    label::Symbol
)

Add an :initial, :final, :control, :state or :variable box constraint on a range.

Note

The range of the constraint must be contained in 1:n if the constraint is on the state, or 1:m if the constraint is on the control, or 1:q if the constraint is on the variable.

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

Examples

julia> constraint!(ocp, :initial, 2:3, [ 0, 0 ], [ 1, 2 ])
julia> constraint!(ocp, :final, Index(1), 0, 2)
julia> constraint!(ocp, :control, Index(1), 0, 2)
julia> constraint!(ocp, :state, 2:3, [ 0, 0 ], [ 1, 2 ])
julia> constraint!(ocp, :initial, 1:2:5, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :variable, 1:2, [ 0, 0 ], [ 1, 2 ])

CTBase.constraint! — Method

constraint!(
    ocp::OptimalControlModel{T, V},
    type::Symbol,
    f::Function,
    lb::Union{Real, AbstractVector{<:Real}},
    ub::Union{Real, AbstractVector{<:Real}}
)
constraint!(
    ocp::OptimalControlModel{T, V},
    type::Symbol,
    f::Function,
    lb::Union{Real, AbstractVector{<:Real}},
    ub::Union{Real, AbstractVector{<:Real}},
    label::Symbol
)

Add a :boundary, :control, :state, :mixed or :variable box functional constraint.

Note

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

When an element is of dimension 1, consider it as a scalar.

Examples

# variable independent ocp
julia> constraint!(ocp, :boundary, (x0, xf) -> x0[3]+xf[2], 0, 1)

# variable dependent ocp
julia> constraint!(ocp, :boundary, (x0, xf, v) -> x0[3]+xf[2]*v[1], 0, 1)

# time independent and variable independent ocp
julia> constraint!(ocp, :control, u -> 2u, 0, 1)
julia> constraint!(ocp, :state, x -> x-1, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :mixed, (x, u) -> x[1]-u, 0, 1)

# time dependent and variable independent ocp
julia> constraint!(ocp, :control, (t, u) -> 2u, 0, 1)
julia> constraint!(ocp, :state, (t, x) -> x-t, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :mixed, (t, x, u) -> x[1]-u, 0, 1)

# time independent and variable dependent ocp
julia> constraint!(ocp, :control, (u, v) -> 2u*v[1], 0, 1)
julia> constraint!(ocp, :state, (x, v) -> x-v[1], [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :mixed, (x, u, v) -> x[1]-v[2]*u, 0, 1)

# time dependent and variable dependent ocp
julia> constraint!(ocp, :control, (t, u, v) -> 2u+v[2], 0, 1)
julia> constraint!(ocp, :state, (t, x, v) -> x-t*v[1], [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :mixed, (t, x, u, v) -> x[1]*v[2]-u, 0, 1)

CTBase.constraint — Method

constraint(
    ocp::OptimalControlModel{T<:TimeDependence, V<:VariableDependence},
    label::Symbol
) -> Any

Retrieve a labeled constraint. The result is a function associated with the constraint computation (not taking into account provided value / bounds).

Example

julia> constraint!(ocp, :initial, 0, :c0)
julia> c = constraint(ocp, :c0)
julia> c(1)
1

CTBase.constraints_labels — Method

constraints_labels(
    ocp::OptimalControlModel
) -> Base.KeySet{Symbol, Dict{Symbol, Tuple}}

Return the labels of the constraints as a Base.keys.

Example

julia> constraints_labels(ocp)

CTBase.control! — Function

control!(ocp::OptimalControlModel, m::Integer)
control!(ocp::OptimalControlModel, m::Integer, name::String)
control!(
    ocp::OptimalControlModel,
    m::Integer,
    name::String,
    components_names::Vector{String}
)

Define the control dimension and possibly the names of each coordinate.

Note

You must use control! only once to set the control dimension.

Examples

julia> control!(ocp, 1)
julia> ocp.control_dimension
1
julia> ocp.control_components_names
["u"]

julia> control!(ocp, 1, "v")
julia> ocp.control_dimension
1
julia> ocp.control_components_names
["v"]

julia> control!(ocp, 2)
julia> ocp.control_dimension
2
julia> ocp.control_components_names
["u₁", "u₂"]

julia> control!(ocp, 2, :v)
julia> ocp.control_dimension
2
julia> ocp.control_components_names
["v₁", "v₂"]

julia> control!(ocp, 2, "v")
julia> ocp.control_dimension
2
julia> ocp.control_components_names
["v₁", "v₂"]

CTBase.dynamics! — Method

dynamics!(
    ocp::OptimalControlModel{T<:TimeDependence, V<:VariableDependence},
    f::Function
)

Set the dynamics.

Note

You can use dynamics! only once to define the dynamics.

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

When an element is of dimension 1, consider it as a scalar.

Example

julia> dynamics!(ocp, f)

CTBase.is_max — Method

is_max(ocp::OptimalControlModel) -> Bool

Return true if the criterion type of ocp is :max.

CTBase.is_min — Method

is_min(ocp::OptimalControlModel) -> Bool

Return true if the criterion type of ocp is :min.

CTBase.is_time_dependent — Method

is_time_dependent(
    ocp::OptimalControlModel{NonAutonomous}
) -> Bool

Return true if the model has been defined as time dependent.

CTBase.is_time_independent — Method

is_time_independent(ocp::OptimalControlModel) -> Bool

Return true if the model has been defined as time independent.

CTBase.is_variable_dependent — Method

is_variable_dependent(
    ocp::OptimalControlModel{<:TimeDependence, NonFixed}
) -> Bool

Return true if the model has been defined as variable dependent.

CTBase.is_variable_independent — Method

is_variable_independent(ocp::OptimalControlModel) -> Bool

Return true if the model has been defined as variable independent.

CTBase.nlp_constraints — Method

nlp_constraints(
    ocp::OptimalControlModel
) -> Tuple{Tuple{Vector{Real}, CTBase.var"#ξ#99", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#η#100", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#ψ#101", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#ϕ#102", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#θ#103", Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}}

Return a 6-tuple of tuples:

(ξl, ξ, ξu) are control constraints

(ηl, η, ηu) are state constraints

(ψl, ψ, ψu) are mixed constraints

(ϕl, ϕ, ϕu) are boundary constraints

(θl, θ, θu) are variable constraints

(ul, uind, uu) are control linear constraints of a subset of indices

(xl, xind, xu) are state linear constraints of a subset of indices

(vl, vind, vu) are variable linear constraints of a subset of indices

Note

The dimensions of the state and control must be set before calling nlp_constraints.

Example

julia> (ξl, ξ, ξu), (ηl, η, ηu), (ψl, ψ, ψu), (ϕl, ϕ, ϕu), (θl, θ, θu),
    (ul, uind, uu), (xl, xind, xu), (vl, vind, vu) = nlp_constraints(ocp)

CTBase.objective! — Method

objective!(
    ocp::OptimalControlModel{T<:TimeDependence, V<:VariableDependence},
    type::Symbol,
    g::Function,
    f⁰::Function
)
objective!(
    ocp::OptimalControlModel{T<:TimeDependence, V<:VariableDependence},
    type::Symbol,
    g::Function,
    f⁰::Function,
    criterion::Symbol
)

Set the criterion to the function g and f⁰. Type can be :bolza. Criterion is :min or :max.

Note

You can use objective! only once to define the objective.

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

When an element is of dimension 1, consider it as a scalar.

Example

julia> objective!(ocp, :bolza, (x0, xf) -> x0[1] + xf[2], (x, u) -> x[1]^2 + u^2) # the control is of dimension 1

CTBase.objective! — Method

objective!(
    ocp::OptimalControlModel{T<:TimeDependence, V<:VariableDependence},
    type::Symbol,
    f::Function
)
objective!(
    ocp::OptimalControlModel{T<:TimeDependence, V<:VariableDependence},
    type::Symbol,
    f::Function,
    criterion::Symbol
)

Set the criterion to the function f. Type can be :mayer or :lagrange. Criterion is :min or :max.

Note

You can use objective! only once to define the objective.

The state, control and variable dimensions must be set before. Use state!, control! and variable!.

The times must be set before. Use time!.

When an element is of dimension 1, consider it as a scalar.

Examples

julia> objective!(ocp, :mayer, (x0, xf) -> x0[1] + xf[2])
julia> objective!(ocp, :lagrange, (x, u) -> x[1]^2 + u^2) # the control is of dimension 1

Warning

If you set twice the objective, only the last one will be taken into account.

CTBase.remove_constraint! — Method

remove_constraint!(ocp::OptimalControlModel, label::Symbol)

Remove a labeled constraint.

Example

julia> remove_constraint!(ocp, :con)

CTBase.state! — Function

state!(ocp::OptimalControlModel, n::Integer)
state!(ocp::OptimalControlModel, n::Integer, name::String)
state!(
    ocp::OptimalControlModel,
    n::Integer,
    name::String,
    components_names::Vector{String}
)

Define the state dimension and possibly the names of each component.

Note

You must use state! only once to set the state dimension.

Examples

julia> state!(ocp, 1)
julia> ocp.state_dimension
1
julia> ocp.state_components_names
["x"]

julia> state!(ocp, 1, "y")
julia> ocp.state_dimension
1
julia> ocp.state_components_names
["y"]

julia> state!(ocp, 2)
julia> ocp.state_dimension
2
julia> ocp.state_components_names
["x₁", "x₂"]

julia> state!(ocp, 2, :y)
julia> ocp.state_dimension
2
julia> ocp.state_components_names
["y₁", "y₂"]

julia> state!(ocp, 2, "y")
julia> ocp.state_dimension
2
julia> ocp.state_components_names
["y₁", "y₂"]

CTBase.time! — Function

time!(
    ocp::OptimalControlModel{<:TimeDependence, NonFixed},
    t0::Real,
    indf::Index
)
time!(
    ocp::OptimalControlModel{<:TimeDependence, NonFixed},
    t0::Real,
    indf::Index,
    name::String
)

Fix initial time, final time is free and given by the variable at the provided index.

Note

You must use time! only once to set either the initial or the final time, or both.

Examples

julia> time!(ocp, 0, Index(2), "t")

CTBase.time! — Function

time!(ocp::OptimalControlModel, t0::Real, tf::Real)
time!(
    ocp::OptimalControlModel,
    t0::Real,
    tf::Real,
    name::String
)

Fix initial and final times to times[1] and times[2], respectively.

Examples

julia> time!(ocp, 0, 1)
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"t"

julia> time!(ocp, 0, 1, "s")
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"s"

julia> time!(ocp, 0, 1, :s)
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"s"

CTBase.time! — Function

time!(
    ocp::OptimalControlModel,
    times::AbstractVector{<:Real}
) -> Any
time!(
    ocp::OptimalControlModel,
    times::AbstractVector{<:Real},
    name::String
) -> Any

Fix initial and final times to times[1] and times[2], respectively.

Examples

julia> time!(ocp, [ 0, 1 ])
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"t"

julia> time!(ocp, [ 0, 1 ], "s")
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"s"

julia> time!(ocp, [ 0, 1 ], :s)
julia> ocp.initial_time
0
julia> ocp.final_time
1
julia> ocp.time_name
"s"

CTBase.time! — Function

time!(
    ocp::OptimalControlModel{<:TimeDependence, NonFixed},
    ind0::Index,
    indf::Index
)
time!(
    ocp::OptimalControlModel{<:TimeDependence, NonFixed},
    ind0::Index,
    indf::Index,
    name::String
)

Initial and final times are free and given by the variable at the provided indices.

Examples

julia> time!(ocp, Index(2), Index(3), "t")

CTBase.time! — Function

time!(
    ocp::OptimalControlModel{<:TimeDependence, NonFixed},
    ind0::Index,
    tf::Real
)
time!(
    ocp::OptimalControlModel{<:TimeDependence, NonFixed},
    ind0::Index,
    tf::Real,
    name::String
)

Fix final time, initial time is free and given by the variable at the provided index.

Examples

julia> time!(ocp, Index(2), 1, "t")

CTBase.variable! — Function

variable!(ocp::OptimalControlModel, q::Integer)
variable!(
    ocp::OptimalControlModel,
    q::Integer,
    name::String
)
variable!(
    ocp::OptimalControlModel,
    q::Integer,
    name::String,
    components_names::Vector{String}
)

Define the variable dimension and possibly the names of each component.

Note

You can use variable! once to set the variable dimension when the model is NonFixed.

Examples

julia> variable!(ocp, 1, "v")
julia> variable!(ocp, 2, "v", [ "v₁", "v₂" ])