Model

CTBase.Index — Type

mutable struct Index

Fields

val::Integer

CTBase.OptimalControlModel — Type

mutable struct OptimalControlModel{time_dependence, dimension_usage} <: CTBase.AbstractOptimalControlModel

Fields

initial_time::Union{Nothing, Real}: Default: nothing

final_time::Union{Nothing, Real}: Default: nothing

time_name::Union{Nothing, String}: Default: nothing

lagrange::Union{Nothing, LagrangeFunction{time_dependence, dimension_usage}} where {time_dependence, dimension_usage}: Default: nothing

mayer::Union{Nothing, MayerFunction{dimension_usage}} where dimension_usage: Default: nothing

criterion::Union{Nothing, Symbol}: Default: nothing

dynamics::Union{Nothing, DynamicsFunction{time_dependence, dimension_usage}} where {time_dependence, dimension_usage}: Default: nothing

dynamics!::Union{Nothing, Function}: Default: nothing

state_dimension::Union{Nothing, Integer}: Default: nothing

state_names::Union{Nothing, Vector{String}}: Default: nothing

control_dimension::Union{Nothing, Integer}: Default: nothing

control_names::Union{Nothing, Vector{String}}: Default: nothing

constraints::Dict{Symbol, Tuple}: Default: Dict{Symbol, Tuple{Vararg{Any}}}()

CTBase.Model — Method

Model(
;
    time_dependence,
    dimension_usage
) -> OptimalControlModel{:autonomous, :scalar}

Returns a new OptimalControlModel instance, that is a model of an optimal control problem.

The model is defined by the following arguments:

time_dependence: either :autonomous or :nonautonomous. Default is :autonomous.

dimension_usage: either :scalar or :vectorial. Default is :scalar.

Examples

julia> ocp = Model()
julia> ocp = Model(time_dependence=:nonautonomous)
julia> ocp = Model(dimension_usage=:vectorial)
julia> ocp = Model(time_dependence=:nonautonomous, dimension_usage=:vectorial)

Note

If the time dependence of the model is defined as nonautonomous, then, the dynamics function, the lagrange cost and the path constraints must be defined as functions of time and state, and possibly control. If the model is defined as autonomous, then, the dynamics function, the lagrange cost and the path constraints must be defined as functions of state, and possibly control.

If the dimension usage of the model is defined as vectorial, then, one dimensional state and / or control variables are considered as vectors. If the model is defined as scalar, then, one dimensional state and / or control variables are considered as scalars.

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    lb,
    ub
) -> Tuple{Symbol, Symbol, CTBase.var"#59#60", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#61#62", Any, Any}

Add an :initial, :final, :control or :state box constraint (whole range).

Examples

julia> constraint!(ocp, :initial, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :final, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :control, [ 0, 0 ], [ 2, 3 ])
julia> constraint!(ocp, :state, [ 0, 0, 0 ], [ 1, 2, 1 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    val
) -> Tuple{Symbol, Symbol, Function, Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, Function, Any, Any}

Add a :boundary, :control, :state or :mixed value functional constraint.

Examples

julia> constraint!(ocp, :boundary, f, [ 0, 0 ])
julia> constraint!(ocp, :control, f, [ 0, 0 ])
julia> constraint!(ocp, :state, f, [ 0, 0 ])
julia> constraint!(ocp, :mixed, f, [ 0, 0 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, UnitRange{<:Integer}},
    val
) -> Tuple{Symbol, Symbol, CTBase.var"#61#62", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, UnitRange{<:Integer}},
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#61#62", Any, Any}

Add an :initial or :final value constraint on a range of the state.

Examples

julia> constraint!(ocp, :initial, 2:3, [ 0, 0 ])
julia> constraint!(ocp, :final, Index(1), 0)

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    lb,
    ub
) -> Tuple{Symbol, Symbol, Function, Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, Function, Any, Any}

Add a :boundary, :control, :state or :mixed box functional constraint.

Examples

julia> constraint!(ocp, :boundary, f, [ 0, 0 ], [ 1, 2 ])
julia> constraint!(ocp, :control, f, [ 0, 0 ], [ 1, 2 ])
julia> constraint!(ocp, :state, f, [ 0, 0 ], [ 1, 2 ])
julia> constraint!(ocp, :mixed, f, [ 0, 0 ], [ 1, 2 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    val
) -> Any
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#59#60", Any, Any}

Add an :initial or :final value constraint on the state.

Examples

julia> constraint!(ocp, :initial, [ 0, 0, 0 ])
julia> constraint!(ocp, :final, [ 0, 0, 0 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, UnitRange{<:Integer}},
    lb,
    ub
) -> Tuple{Symbol, Symbol, CTBase.var"#61#62", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, UnitRange{<:Integer}},
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, Any, Any, Any}

Add an :initial, :final, :control or :state box constraint on a range.

Examples

julia> constraint!(ocp, :initial, 2:3, [ 0, 0 ], [1, 2])
julia> constraint!(ocp, :final, Index(1), 0, 2)
julia> constraint!(ocp, :control, Index(1), 0, 2)
julia> constraint!(ocp, :state, 2:3, [ 0, 0 ], [1, 2])

CTBase.constraint! — Method

constraint!(
    ocp::OptimalControlModel{time_dependence, dimension_usage},
    type::Symbol,
    f::Function
) -> Any

Provide dynamics (possibly in place).

Examples

julia> constraint!(ocp, :dynamics, f)
julia> constraint!(ocp, :dynamics!, f!)

CTBase.constraint — Method

constraint(
    ocp::OptimalControlModel,
    label::Symbol
) -> Function

Retrieve a labeled constraint. The result is a function associated with the constraint computation (not taking into account provided value / bounds).

Examples

julia> constraint(ocp, :eq)

CTBase.constraints_labels — Method

constraints_labels(
    ocp::OptimalControlModel
) -> Base.KeySet{Symbol, Dict{Symbol, Tuple}}

Returns the labels of the constraints as a Base.keys.

Examples

julia> constraints_labels(ocp)

CTBase.control! — Function

control!(
    ocp::OptimalControlModel,
    m::Integer
) -> Union{Vector{Vector{String}}, Vector{String}, String}
control!(
    ocp::OptimalControlModel,
    m::Integer,
    names::Union{String, Vector{String}}
) -> Union{Vector{Vector{String}}, Vector{String}, String}

Define the control dimension and possibly the names of each coordinate.

Examples

julia> control!(ocp, 1, "v")
julia> control!(ocp, 2, [ "u₁", "u₂" ])

CTBase.isautonomous — Method

isautonomous(ocp::OptimalControlModel) -> Bool

Returns true if the model has been defined as autonomous.

CTBase.isautonomous — Method

isautonomous(time_dependence::Symbol) -> Bool

Returns !isnonautonomous(time_dependence)

CTBase.ismax — Method

ismax(ocp::OptimalControlModel) -> Bool

Returns true if the criterion type of ocp is :max.

CTBase.ismin — Method

ismin(ocp::OptimalControlModel) -> Bool

Returns true if the criterion type of ocp is :min.

CTBase.isnonautonomous — Method

isnonautonomous(time_dependence::Symbol) -> Bool

Returns :nonautonomous == time_dependence

CTBase.isnonautonomous — Method

isnonautonomous(
    ocp::OptimalControlModel{time_dependence, dimension_usage}
) -> Bool

Returns true if the model has been defined as nonautonomous.

CTBase.nlp_constraints — Method

nlp_constraints(
    ocp::OptimalControlModel{time_dependence, dimension_usage}
) -> Tuple{Tuple{Vector{Real}, CTBase.var"#ξ#77", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#η#78", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#ψ#79", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#ϕ#80", Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}}

Return a 6-tuple of tuples:

(ξl, ξ, ξu) are control constraints

(ηl, η, ηu) are state constraints

(ψl, ψ, ψu) are mixed constraints

(ϕl, ϕ, ϕu) are boundary constraints

(ulb, uind, uub) are control linear constraints of a subset of indices

(xlb, xind, xub) are state linear constraints of a subset of indices

Examples

julia> (ξl, ξ, ξu), (ηl, η, ηu), (ψl, ψ, ψu), (ϕl, ϕ, ϕu), 
    (ulb, uind, uub), (xlb, xind, xub) = nlp_constraints(ocp)

CTBase.objective! — Method

objective!(
    ocp::OptimalControlModel{time_dependence, dimension_usage},
    type::Symbol,
    g::Function,
    f⁰::Function
) -> Any
objective!(
    ocp::OptimalControlModel{time_dependence, dimension_usage},
    type::Symbol,
    g::Function,
    f⁰::Function,
    criterion::Symbol
) -> Any

Set the criterion to the function g and f⁰. Type can be :bolza. Criterion is :min or :max.

Examples

julia> objective!(ocp, :bolza, (t0, x0, tf, xf) -> tf, (x, u) -> x[1]^2 + u[1]^2) 

CTBase.objective! — Method

objective!(
    ocp::OptimalControlModel{time_dependence, dimension_usage},
    type::Symbol,
    f::Function
) -> Any
objective!(
    ocp::OptimalControlModel{time_dependence, dimension_usage},
    type::Symbol,
    f::Function,
    criterion::Symbol
) -> Any

Set the criterion to the function f. Type can be :mayer or :lagrange. Criterion is :min or :max.

Examples

julia> objective!(ocp, :mayer, (t0, x0, tf, xf) -> tf)
julia> objective!(ocp, :lagrange, (x, u) -> x[1]^2 + u[1]^2)

Important

If you set twice the objective, only the last one will be taken into account.

CTBase.remove_constraint! — Method

remove_constraint!(
    ocp::OptimalControlModel,
    label::Symbol
) -> Dict{Symbol, Tuple}

Remove a labeled constraint.

Examples

julia> remove_constraint!(ocp, :con)

CTBase.state! — Function

state!(
    ocp::OptimalControlModel,
    n::Integer
) -> Union{Vector{Vector{String}}, Vector{String}, String}
state!(
    ocp::OptimalControlModel,
    n::Integer,
    names::Union{String, Vector{String}}
) -> Union{Vector{Vector{String}}, Vector{String}, String}

Define the state dimension and possibly the names of each coordinate.

Examples

julia> state!(ocp, 1, "y")
julia> state!(ocp, 2, [ "x₁", "x₂" ])

CTBase.time! — Function

time!(
    ocp::OptimalControlModel,
    times::AbstractVector{<:Real}
) -> String
time!(
    ocp::OptimalControlModel,
    times::AbstractVector{<:Real},
    name::String
) -> String

Fix initial and final times to t[1] and t[2], respectively.

Examples

julia> time!(ocp, [ 0, 1 ], "t")

CTBase.time! — Function

time!(
    ocp::OptimalControlModel,
    type::Symbol,
    time::Real
) -> String
time!(
    ocp::OptimalControlModel,
    type::Symbol,
    time::Real,
    name::String
) -> String

Fix initial (resp. final) time, the final (resp. initial) time being variable (free), when type is :initial. And conversely when type is :final.

Examples

julia> time!(ocp, :initial, 0, "t")
julia> time!(ocp, :final, 1, "t")