Plot

RecipesBase.plot — Method

plot(
    sol::OptimalControlSolution;
    layout,
    state_style,
    control_style,
    adjoint_style,
    kwargs...
) -> Any

Plot the optimal control solution sol using the layout layout. The argument layout can be :group or :split (default).

Note

The keyword arguments state_style, control_style and adjoint_style are passed to the plot function of the Plots package. The state_style is passed to the plot of the state, the control_style is passed to the plot of the control and the adjoint_style is passed to the plot of the adjoint.

Let us consider we have defined an optimal control problem

ocp = Model()
...

and solve it

sol = solve(ocp)

We can plot the solution with the default layout :split.

plot(sol, layout=:split, size=(800, 600))

Or with the layout :group.

plot(sol, layout=:group, size=(800, 300))

You can specify some styles:

plot(sol,
    state_style=(color=:blue,),
    adjoint_style=(color=:black, linestyle=:dash),
    control_style=(color=:red, linewidth=2),
    size=(800, 600))