Plot

RecipesBase.plot! — Method

plot!(
    p::Plots.Plot,
    sol::OptimalControlSolution;
    layout,
    control,
    state_style,
    control_style,
    costate_style,
    kwargs...
) -> Plots.Plot

Plot the optimal control solution sol using the layout layout.

Notes.

The argument layout can be :group or :split (default).

The keyword arguments state_style, control_style and costate_style are passed to the plot function of the Plots package. The state_style is passed to the plot of the state, the control_style is passed to the plot of the control and the costate_style is passed to the plot of the costate.

RecipesBase.plot — Method

plot(
    sol::OptimalControlSolution;
    layout,
    state_style,
    control_style,
    costate_style,
    kwargs...
) -> Any

Plot the optimal control solution sol using the layout layout.

Notes.

The argument layout can be :group or :split (default).

The keyword arguments state_style, control_style and costate_style are passed to the plot function of the Plots package. The state_style is passed to the plot of the state, the control_style is passed to the plot of the control and the costate_style is passed to the plot of the costate.

Let us consider we have defined an optimal control problem

ocp = Model()
...

and solve it

sol = solve(ocp)

We can plot the solution with the default layout :split.

plot(sol, layout=:split, size=(800, 600))

Or with the layout :group.

plot(sol, layout=:group, size=(800, 300))

You can specify some styles:

plot(sol,
    state_style=(color=:blue,),
    costate_style=(color=:black, linestyle=:dash),
    control_style=(color=:red, linewidth=2),
    size=(800, 600))

You can also add additional plots with the command plot!.